chore: update to SpringBoot 3.5, introduce ActivityParticipantRepository to troubleshoot issue introduced by updated SpringDataJpa, group Repository tests in RepositoryITTest
