+++ src/main/java/org/montrealjug/billetterie/BilletterieApplication.java
@@ -5,9 +5,11 @@ import org.montrealjug.billetterie.config.BilletterieProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @EnableConfigurationProperties(BilletterieProperties.class)
+@EnableJpaRepositories(basePackages = "org.montrealjug.billetterie.repository")
 public class BilletterieApplication {
 
     public static void main(String[] args) {
